Collectible turn on/off when player close & added GameManager
