+++ Assets/Prefabs/Collectible.prefab.meta
