+++ Assets/Scenes/SampleScene.unity
@@ -296,72 +296,3 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &2250175202406421162
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 2250175203114752824, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_Name
-      value: BreadCrumbs
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 2
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 2250175203114752828, guid: 1b82de60fab5eb8419bef751243c92aa,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 1b82de60fab5eb8419bef751243c92aa, type: 3}
