+++ Assets/Scripts/Collectibles.meta
