+++ Assets/Scripts/Collectibles/AppearToPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b705def71392ff54ba6e255f4044b0d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
