+++ Assets/Scripts/Collectibles/CollectibleBehaviour.cs.meta
