+++ ProjectSettings/TagManager.asset
@@ -13,8 +13,8 @@ TagManager:
   - UI
   - 
   - 
-  - 
-  - 
+  - Collectible
+  - Player
   - 
   - 
   -
